revert to python 2.7

+++ scripts/models.py
@@ -11,7 +11,7 @@ from glob import glob
 
 def Deconv(inputs, f_dim_in, dim, net, batch_size, f_dim_out = None, stride = 2 ):
 	if f_dim_out is None: 
-		f_dim_out = int(f_dim_in/2)
+		f_dim_out = f_dim_in/2 
 	return tl.layers.DeConv3dLayer(inputs,
 								shape = [4, 4, 4, f_dim_out, f_dim_in],
 								output_shape = [batch_size, dim, dim, dim, f_dim_out],
